requirements.txt
ta==0.10.1
tqdm==4.64.0
pandas_ta==0.3.14b0
pytorch==1.13.1
torchvision==0.14.1
requests==2.28.2
tqdm==4.65.0